api/requirements.txt
fastapi ==0.88.0
uvicorn == 0.20.0
passlib[bcrypt] == 1.7.4
pymysql == 1.0.2
pydantic
python-multipart
pandas
cryptography ==38.0.4
bs4
requests
asyncio
aiohttp
chardet